refactor/PK-16: Modal Base 컴포넌트 구현
- 함께패킹리스트 모달을 Modal 컴포넌트 합성으로 구조 변경
