+++ components/common/Modal.tsx
@@ -1,75 +1,109 @@
+import React, { PropsWithChildren, ReactElement } from 'react';
+import ReactDOM from 'react-dom';
+import Image from 'next/image';
 import styled from 'styled-components';
 import iClose from '/public/assets/svg/iClose.svg';
-import Image from 'next/image';
-import { FONT_STYLES } from '../../styles/font';
+import { packmanColors } from '../../styles/color';
 
 interface ModalProps {
-  title: string;
-  button?: React.ReactNode;
-  closeModal: () => void;
+  onClick: VoidFunction;
+  closeBtn?: boolean;
 }
-export default function Modal(props: ModalProps) {
-  const { title, button, closeModal } = props;
 
-  return (
-    <StyledBackground>
-      <StyledRoot>
-        <StyledImageWrapper>
-          <Image src={iClose} alt="close" onClick={closeModal} />
-        </StyledImageWrapper>
-        <StyledModalInfo>
-          <p>{title}</p>
-          {button}
-        </StyledModalInfo>
-      </StyledRoot>
-    </StyledBackground>
+function Modal({ children, onClick, closeBtn = true }: PropsWithChildren<ModalProps>) {
+  return ReactDOM.createPortal(
+    <>
+      <Modal.Bg onClick={onClick}></Modal.Bg>
+      <StyledModal>
+        {closeBtn && <Modal.CloseBtn onClick={onClick} />}
+        {children}
+      </StyledModal>
+    </>,
+    document.querySelector('#__next') as Element,
   );
 }
-const StyledBackground = styled.div`
+
+type BgProps = ModalProps;
+
+Modal.Bg = function Bg({ onClick }: BgProps) {
+  return <StyledBg onClick={onClick} />;
+};
+
+type CloseBtnProps = ModalProps;
+
+Modal.CloseBtn = function CloseBtn({ onClick }: CloseBtnProps) {
+  return (
+    <StyledButtonContainer>
+      <StyledCloseBtn onClick={onClick}>
+        <Image src={iClose} alt="closeModal" width="24" height="24" />
+      </StyledCloseBtn>
+    </StyledButtonContainer>
+  );
+};
+
+Modal.Body = function Body({ children }: PropsWithChildren) {
+  return <StyledBody>{children}</StyledBody>;
+};
+
+interface ControlsProps {
+  controls: ReactElement;
+}
+
+Modal.Controls = function Controls({ controls }: ControlsProps) {
+  return controls;
+};
+
+export default Modal;
+
+const StyledBg = styled.div`
   position: fixed;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.47);
-  z-index: 110;
-  overflow: hidden;
-`;
-const StyledRoot = styled.div`
+  width: 100vw;
+  height: calc(var(--vh, 1vh) * 100);
+  background-color: rgba(0, 0, 0, 0.48);
   display: flex;
   flex-direction: column;
-  justify-content: center;
-  text-align: center;
+  align-items: center;
+  z-index: 9999;
+  left: 0;
+  top: 0;
+`;
 
-  position: absolute;
+const StyledModal = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  position: fixed;
   top: 50%;
   left: 50%;
+  z-index: 99999;
   transform: translate(-50%, -50%);
+
   width: 32rem;
-  height: 17.6rem;
-  background-color: #fff;
-  border-radius: 1.34rem;
 
-  z-index: 115;
+  background-color: ${packmanColors.pmWhite};
 
-  & > span {
-    position: absolute;
-    top: 0.7rem;
-    right: 1.5rem;
-  }
+  padding: 0.8rem 1.6rem 2.4rem 1.6rem;
+
+  border-radius: 1rem;
 `;
-const StyledImageWrapper = styled.div`
-  position: absolute;
-  top: 0.7rem;
-  right: 1.4rem;
-`;
-const StyledModalInfo = styled.div`
+
+const StyledButtonContainer = styled.div`
+  width: 100%;
   display: flex;
-  flex-direction: column;
-  justify-content: center;
+  flex-direction: row-reverse;
+  margin-right: -1.6rem;
+  margin-bottom: 1.1rem;
+`;
 
-  & > p {
-    ${FONT_STYLES.SUBHEAD1_SEMIBOLD};
-    padding-bottom: 3.7rem;
-    color: #282828;
-  }
+const StyledCloseBtn = styled.button`
+  height: 2.4rem;
+
+  background: ${packmanColors.pmWhite};
+
+  border: none;
+`;
+
+const StyledBody = styled.div`
+  width: 100%;
 `;
